+++ src/components/header/MainHeader.tsx
@@ -1,17 +1,22 @@
+'use client'
 import Image from 'next/image';
-import React from 'react';
 import momodoLogo from '../../../public/images/momodo.svg';
-import calendarIcon from '../../../public/images/calendarIcon.svg';
 import logoutIcon from '../../../public/images/logoutIcon.svg';
+import useModal from '../../hooks/useModal';
+import InnerModal from '../Modal/InnerModal';
 
 const MainHeader: React.FC = () => {
+  const { modalOpen, openModal, closeModal } = useModal();
   return (
-    <header className='flex justify-center relative mb-10'>
-      <Image src={momodoLogo} alt='모모두' width={73} className='' />
-      <button className='absolute right-0 top-0'>
-        <Image src={logoutIcon} alt='로그아웃 버튼' />
-      </button>
-    </header>
+    <>
+      <header className='flex justify-center relative mb-10'>
+        <Image src={momodoLogo} alt='모모두' width={73} className='' />
+        <button className='absolute right-0 top-0' onClick={openModal}>
+          <Image src={logoutIcon} alt='로그아웃 버튼' />
+        </button>
+      </header>
+      {modalOpen && <InnerModal type={'logout'} id={0} closeModal={closeModal} />}
+    </>
   );
 };
