+++ src/hooks/useModal.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { useState } from 'react';
 
 const useModal = () => {
